Move structs and corresponding fuctions out of lib.rs
DM, DevId -> dm.rs
Device -> device.rs
DeviceInfo -> deviceinfo.rs

Update use statements in all files to reflect the changes.
