+++ src/lineardev.rs
@@ -6,7 +6,9 @@ use std::fmt;
 use std::fs::File;
 use std::path::PathBuf;
 
-use {DM, DevId, DeviceInfo, DmFlags};
+use consts::DmFlags;
+use deviceinfo::DeviceInfo;
+use dm::{DM, DevId};
 use result::{DmResult, DmError, InternalError};
 use segment::Segment;
 use types::{Bytes, Sectors, TargetLine};
