+++ src/thinpooldev.rs
@@ -5,7 +5,9 @@
 use std::fmt;
 use std::path::PathBuf;
 
-use {DM, DevId, DeviceInfo, DmFlags};
+use consts::DmFlags;
+use deviceinfo::DeviceInfo;
+use dm::{DM, DevId};
 use lineardev::LinearDev;
 use result::{DmResult, DmError, InternalError};
 use types::{DataBlocks, Sectors, TargetLine};
@@ -89,10 +91,10 @@ impl ThinPoolDev {
 
         DM::wait_for_dm();
         Ok(ThinPoolDev {
-               dev_info: di,
-               meta_dev: meta,
-               data_dev: data,
-           })
+            dev_info: di,
+            meta_dev: meta,
+            data_dev: data,
+        })
     }
 
     /// Generate a Vec<> to be passed to DM. The format of the Vec
